+++ stdlib/src/stdlib.c
@@ -46,5 +46,7 @@ void *dt_memset(void *dest, int value, size_t n) {
 }
 
 int dt_abs(int value) { return abs(value); }
-double dt_log(double x) { return log(x); }
-double dt_exp(double x) { return exp(x); }
+/* Single-precision logf/expf, not log/exp: on a 32-bit MCU with a single-
+ * precision FPU the double versions are software-emulated and far slower. */
+float dt_log(float x) { return logf(x); }
+float dt_exp(float x) { return expf(x); }
